corretto file migration

+++ database/migrations/2021_07_15_123751_create_travels_table.php
@@ -18,8 +18,8 @@ class CreateTravelsTable extends Migration
             $table->string('area', 25);
             $table->string('country', 25);
             $table->string('name', 100);
-            $table->unsignedSmallInteger('duration_day');
-            $table->float('price_eur', 7, 2); // 99.999,99
+            $table->unsignedTinyInteger('duration_day');
+            $table->float('price_eur', 7, 2);
             $table->string('type', 100);
             $table->string('period', 15)->default('Tutto l\'anno');
             $table->text('info');
